Fix angular sum loop to actually sum angles
Turns out *this* was the cause of the buggy behavior I was seeing
in the last commit: there was really only one "fulcrum", which was
jumping when that fulcrum changed in sort order.

The bug I described is still a problem if you want deltas to
reflect actual changes in degree, but it yields correct values when
mapping onto a rotation.